memberpage-20220519

+++ Convenience_POS_System/src/main/java/com/example/convenience_pos_system/dao/ProductHistoryDao.java
@@ -72,7 +72,7 @@ public class ProductHistoryDao {
         return results.isEmpty() ? null : results;
     }
 
-    public List<ProductHistory> selectByMid(int mid) {
+    public List<ProductHistory> selectByMid(Long mid) {
         List<ProductHistory> results = jdbcTemplate.query(
                 "select * from PRODUCT_HISTORY_TB where MID = ?",
                 new RowMapper<ProductHistory>() {
